+++ 2-Core/MoContent3d/Resource/Scene/FDrScenePass.cs
@@ -66,8 +66,8 @@ namespace MO.Content3d.Resource.Scene
       //============================================================
       public void Serialize(IOutput output) {
          output.WriteString(_name);
-         output.WriteInt32(_renderTargetSize.Width);
-         output.WriteInt32(_renderTargetSize.Height);
+         output.WriteUint16((ushort)_renderTargetSize.Width);
+         output.WriteUint16((ushort)_renderTargetSize.Height);
       }
 
       //============================================================
